Add toggle to show all orbit traces
