+++ src/components/FocusControls.tsx
@@ -1,9 +1,9 @@
-import { ActionIcon, Box, Button, Group, Menu, Stack, Text, Tooltip, Transition } from '@mantine/core';
-import { IconCircle, IconCircleDot, IconCircleFilled, IconEyeCog } from '@tabler/icons-react';
+import { Box, Button, Group, Menu, Stack, Text, Transition } from '@mantine/core';
+import { IconCircle, IconCircleDot, IconCircleFilled } from '@tabler/icons-react';
 import { CELESTIAL_BODY_CLASSES, CELESTIAL_BODY_NAMES, findCelestialBody } from '../lib/constants.ts';
 import { celestialBodyTypeName } from '../lib/utils.ts';
 import { AppState } from '../lib/state.ts';
-import { CelestialBodyState, CelestialBodyType } from '../lib/types.ts';
+import { CelestialBodyState } from '../lib/types.ts';
 import { FactCard } from './FactCard.tsx';
 import { useMemo } from 'react';
 
@@ -20,13 +20,6 @@ export function FocusControls({ state, updateState, systemState }: Props) {
     () => (state.hover != null ? findCelestialBody(systemState, state.hover) : null),
     [state.hover]
   );
-
-  function toggleVisibleType(type: CelestialBodyType) {
-    const newVisibleTypes = state.visibleTypes.has(type)
-      ? new Set([...state.visibleTypes].filter(t => t !== type))
-      : new Set([...state.visibleTypes, type]);
-    updateState({ visibleTypes: newVisibleTypes });
-  }
 
   return (
     <Stack gap="xs">
@@ -48,26 +41,6 @@ export function FocusControls({ state, updateState, systemState }: Props) {
                   <Text size="xs" c="dimmed">
                     {celestialBodyTypeName(CELESTIAL_BODY_CLASSES[i])}
                   </Text>
-                </Group>
-              </Menu.Item>
-            ))}
-          </Menu.Dropdown>
-        </Menu>
-
-        <Menu position="top-start" offset={0} width={120} closeOnItemClick={false}>
-          <Menu.Target>
-            <Tooltip position="right" label="Toggle Visibility">
-              <ActionIcon>
-                <IconEyeCog size={iconSize} />
-              </ActionIcon>
-            </Tooltip>
-          </Menu.Target>
-          <Menu.Dropdown {...menuDropdownProps}>
-            {(['sun', 'planet', 'moon', 'asteroid', 'trans-neptunian-object'] as Array<CelestialBodyType>).map(type => (
-              <Menu.Item key={type} onClick={() => toggleVisibleType(type)}>
-                <Group gap="xs" align="center">
-                  {state.visibleTypes.has(type) ? <IconCircleFilled size={14} /> : <IconCircle size={14} />}
-                  {celestialBodyTypeName(type)}
                 </Group>
               </Menu.Item>
             ))}
